Bugfix: Adds support for eureka server to run and publish service through configured environment variable
Even though Eureka server was configured use environment variable for publishing the service.
When ran in docker it was experiencing exception : c.n.e.cluster.ReplicationTaskProcessor   : Network level connection to peer localhost. Coz it was looking for a peer instance on localhost:8761.
This commit, forces it use 1111 port as desired.